move util function to Util
